+++ avr/cores/megacommand/MCL/EmptyTrack.h
@@ -13,10 +13,10 @@ public:
 
   uint8_t data[EMPTY_TRACK_LEN];
         /*
-  bool get_track_from_sysex(int tracknumber, uint8_t column);
-  void place_track_in_sysex(int tracknumber, uint8_t column);
-  bool load_track_from_grid(int32_t column, int32_t row, int m);
-  bool store_track_in_grid(int track, int32_t column, int32_t row);
+  bool get_track_from_sysex(uint8_t tracknumber, uint8_t column);
+  void place_track_in_sysex(uint8_t tracknumber, uint8_t column);
+  bool load_track_from_grid(uint8_t column, uint8_t row, int m);
+  bool store_track_in_grid(uint8_t track, uint8_t column, uint8_t row);
 */
 };
