+++ avr/cores/megacommand/MCL/FileBrowserPage.h
@@ -26,7 +26,7 @@ class FileBrowserPage : public LightPage {
 public:
   File file;
   //  char file_entries[NUM_FILE_ENTRIES][16];
-  int numEntries;
+  int16_t numEntries;
 
   char match[5];
   char lwd[128];
@@ -76,7 +76,7 @@ public:
 private:
 
   void _handle_filemenu();
-  void _calcindices(int &);
+  void _calcindices(int16_t &);
   void _cd_up();
   void _cd(const char *);
 };
